Compile Juniper ConfigurationManagement code first
Unity projects have a weird order in which things get compiled. The
Juniper Unity scripts have a lot of dependencies on other code, which
ideally gets managed by compiler define symbols. But this doesn't
always happen, especially when first setting up a project. So making
sure the config mgmt code gets compiled separately from the rest
of the Unity scripts makes sure that it is available, even if the
rest of the project configuration is "broken".